+++ compiler/src/main/java/info/teksol/mc/profile/GlobalCompilerProfile.java
@@ -69,7 +69,11 @@ public interface GlobalCompilerProfile {
     }
 
     default boolean isSignature() {
-        return getBooleanValue(MlogFormatOptions.SIGNATURE);
+        return getBooleanValue(CompilerOptions.SIGNATURE);
+    }
+
+    default List<String> getAuthors() {
+        return this.<String>getOption(CompilerOptions.AUTHOR).getValues();
     }
 
     default boolean isShortFunctionPrefix() {