+++ compiler/src/main/java/info/teksol/mc/profile/options/CompilerOptions.java
@@ -4,12 +4,14 @@ import org.jspecify.annotations.NullMarked;
 
 @NullMarked
 public enum CompilerOptions {
+    AUTHOR,
     AUTO_PRINTFLUSH,
     BOUNDARY_CHECKS,
     EMULATE_STRICT_NOT_EQUAL,
     ERROR_FUNCTION,
     ERROR_REPORTING,
     REMARKS,
+    SIGNATURE,
     SYNTAX,
     TARGET_GUARD,
 }